🥅 Added error handling

+++ chatgpt/chatgpt.py
@@ -5,7 +5,8 @@ import openai
 from loguru import logger
 from telethon.tl.types import User
 
-from chatgpt.utils import UserType
+from chatgpt.utils import DataType, UserType
+from sqlitedb.utils import ErrorCodes
 
 
 class ChatGPT(object):
@@ -52,49 +53,67 @@ class ChatGPT(object):
         reply = str(response["choices"][0]["message"]["content"])
         return reply
 
-    def chat(self, user: User, message: str) -> str:
+    def chat(self, user: User, message: str) -> str | int:
         """Chat Open API."""
         from main import db
 
-        db.insert_message_from_user(message, user.id)
+        response = db.insert_message_from_user(message, user.id)
+        if response <= ErrorCodes.exceptions.value:
+            return response
         messages = db.get_messages_by_user(user.id)
+        if response <= ErrorCodes.exceptions.value:
+            return response
         self.message_history[user.username] = self.build_message(messages)
-        response = openai.ChatCompletion.create(  # type: ignore
+        openapi_response = openai.ChatCompletion.create(  # type: ignore
             model="gpt-3.5-turbo",
             messages=self.message_history[user.username],
             timeout=10,
         )
         logger.debug("Got response fromm open AI")
-        reply = str(response["choices"][0]["message"]["content"])
-        db.insert_message_from_gpt(reply, user.id)
+        reply = str(openapi_response["choices"][0]["message"]["content"])
+        response = db.insert_message_from_gpt(reply, user.id)
+        if response <= ErrorCodes.exceptions.value:
+            return response
         self.message_history[user.username].append(
             {"role": UserType.ASSISTANT.value, "content": reply}
         )
         return reply
 
-    def image_gen(self, telegram_user: User, message: str) -> str:
+    def image_gen(self, telegram_user: User, message: str) -> str | int:
         """Generate an image from the text."""
         response = openai.Image.create(prompt=message, n=1, size="512x512")  # type: ignore
         image_url = str(response["data"][0]["url"])
         from main import db
 
-        db.insert_images_from_gpt(message, image_url, telegram_user.id)
+        response = db.insert_images_from_gpt(message, image_url, telegram_user.id)
+        if response <= ErrorCodes.exceptions.value:
+            return int(response)
         return image_url
 
-    def clean_up_user_messages(self, telegram_user: User) -> int:
+    def _clean_up_user_messages(self, telegram_user: User) -> int:
         """Delete all user's message data."""
         from main import db
 
         return db.delete_all_user_messages(telegram_user.id)
 
-    def clean_up_user_images(self, user: User) -> int:
+    def _clean_up_user_images(self, user: User) -> int:
         """Delete all user's image data."""
         from main import db
 
-        return db.delete_all_user_messages(user.id)
+        return db.delete_all_user_images(user.id)
 
-    def clean_up_user_data(self, telegram_user: User) -> Tuple[int, int]:
+    def clean_up_user_data(
+        self, data: str, telegram_user: User
+    ) -> Tuple[int, int] | int:
         """Delete all for a user data."""
         from main import db
 
-        return db.delete_all_user_data(telegram_user.id)
+        if data == DataType.MESSAGES.value:
+            return self._clean_up_user_messages(telegram_user)
+        elif data == DataType.IMAGES.value:
+            return self._clean_up_user_images(telegram_user)
+        elif data == DataType.ALL.value:
+            return db.delete_all_user_data(telegram_user.id)
+        else:
+            logger.error("Not a valid choice")
+            return -1, -1
